Update unknown_threat.yara
we had issues :

The string used in the .yara rule is incorrect. Below are few suggestions that could help to make the answer correct.

As you have not identified the "callout domain" in previous tasks, this answer will anyway be incorrect.
One more important thing, a .yara rule string is a well thought, carefully chosen string, so that it not only prevents "this particular variation" of attack but also cover more attack variation in a single signature.
Also the more string or the larger string you set, it will look for exact match and attacker can easily do slight changes to file contents to bypass your .yara signature check.
We should only look for the key data which will not change. Example the "callout domain" only. Generally it is not that easy to change the callout domain as compared to other data. Hence currently this is the only IOC here (indicator of compromise) here.